Removed redundancies
